requirement.txt
absl-py==1.0.0
attrs==21.4.0
autoflake==1.4
black==22.3.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
confluent-kafka==1.8.2
cycler==0.11.0
flake8==4.0.1
flake8-bugbear==22.3.23
flake8-builtins==1.5.3
flake8-comprehensions==3.8.0
flake8-isort==4.1.1
flake8-polyfill==1.0.2
fonttools==4.33.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
isort==5.10.1
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
pep8-naming==0.12.1
Pillow==9.1.0
platformdirs==2.5.1
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
testfixtures==6.18.5
thop==0.0.31.post2005241907
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
websockets==10.3
Werkzeug==2.1.1
